+++ src/main/java/org/example/kinogris/Service/ShowingService.java
@@ -34,6 +34,9 @@ public class ShowingService {
     }
 
     public Optional<Showing> updateShowing(int id, @Valid Showing showingDetails) {
+        if(showingDetails == null){
+            throw new IllegalArgumentException("The showing details cannot be null");
+        }
         return showingRepository.findById(id).map(showing -> {
             showing.setMovie(showingDetails.getMovie());
             showing.setTheater(showingDetails.getTheater());
@@ -45,6 +48,9 @@ public class ShowingService {
     }
 
     public boolean deleteShowing(int id) {
+        if(id <= 0 ){ //checker for en valid id
+            throw new IllegalArgumentException("The showing details cannot be null");
+        }
         if (showingRepository.existsById(id)) {
             showingRepository.deleteById(id);
             return true;
@@ -52,6 +58,9 @@ public class ShowingService {
         return false;
     }
     public boolean isConflict(Showing showing) {
+        if(showing.getMovie() == null || showing.getEndTime() == null){
+            throw new IllegalArgumentException("The showing details cannot be null");
+        }
         List<Showing> existingShowings = showingRepository.findByStartTimeBetween(showing.getStartTime(), showing.getEndTime());
         return !existingShowings.isEmpty();
     }
